requirements.txt
numpy
pandas
python-docx
litellm
anthropic
python-dotenv